:recycle: Refactor Layout
